Extract PDF viewer HTML template into separate module
Move the HTML template, CSS, SVG icons, and escape helper into
functions/_lib/pdf-viewer.js, keeping the route handler focused
on request logic.
